Add files via upload
In this segment I tried bining the data, instead of doing away with the outliers which leads to a loss of almost 300k rows. By so doing we might be loosing a segment in the market that I named luxury. I ended up with 4 bins using the quartile range. This way we still get to keep data that would otherwise be lost.